Refactor: package isolation
