+++ README.md
@@ -105,3 +105,74 @@
 + val value: String? = readLine()
 + val value: String = readLine()!!
 + val number = readLine()!!.toInt()
+
+
+## STEP3 (4단계 - 자동차 경주)
+### 기능 요구 사항
+초간단 자동차 경주 게임을 구현한다.
+
+- 각 자동차에 이름을 부여할 수 있다. 자동차 이름은 5자를 초과할 수 없다.
+  - 자동차 이름이 빈 문자열 일 경우 예외(IllegalArgumentException)를 던진다.
+  - 자동차 이름이 5자리를 초과하는 경우 예외(IllegalArgumentException)를 던진다.
+  - 자동차 이름이 "pobi"로 주어 졌을 경우 "pobi"라는 자동차가 만들어 져야한다. 
+- 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
+- 자동차 이름은 쉼표(,)를 기준으로 구분한다.
+  - 자동차 대수가 1대일 때 쉼표가 있으면 예외(IllegalArgumentException)를 던진다.
+  - 자동차 쉼표가 연속으로 나올 경우 예외(IllegalArgumentException)를 던진다.
+  - 자동차 이름이 "pobi,crong,honux"로 주어졌을 경우 자동차 객체는 3개가 만들어져야아며 각 이름은 "pobi", "crong", "honux"이여야 한다.
+- 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한 명 이상일 수 있다.
+  - 우승자가 한명도 없을 경우 예외(IllegalArgumentException)를 던진다.
+  - 우승자가 한명일 경우 우승자의 숫자가 1이여야 한다.
+  - 우승자가 3명일 경우 우승자의 숫자가 3이여야 한다.
+
+### 기능 목록
+- [ ] 자동차는 이름을 가진다.
+- [ ] 자동차의 이름은 비어 있을 수 없다.
+- [ ] 자동차의 이름은 5자리를 초과할 수 없다.
+- [ ] 자동차의 이름은 쉼표(,)로 구분 된다.
+- [ ] 자동차의 이름 입력 맨 뒤에 쉼표가 올수 없다.
+- [ ] 자동차 경주의 우승자가 1명도 없을 수 없다.
+- [ ] 자동차 경주의 우승자가 여러명 일수 있다.
+
+### 프로그래밍 요구 사항
+- 모든 로직에 단위 테스트를 구현한다. 단, UI(System.out, System.in) 로직은 제외
+- 핵심 로직을 구현하는 코드와 UI를 담당하는 로직을 구분한다.
+- UI 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
+- indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다. 1까지만 허용한다.
+  - 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다.
+  - 힌트: indent(인덴트, 들여쓰기) depth를 줄이는 좋은 방법은 함수(또는 메소드)를 분리하면 된다.
+- 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다.
+- 함수(또는 메소드)가 한 가지 일만 잘 하도록 구현한다.
+- 기능을 구현하기 전에 README.md 파일에 구현할 기능 목록을 정리해 추가한다.
+- git의 commit 단위는 앞 단계에서 README.md 파일에 정리한 기능 목록 단위로 추가한다.
+
+### 실행 결과
+위 요구 사항에 따라 3대의 자동차가 5번 움직였을 경우 프로그램을 실행한 결과는 다음과 같다.
+
+ 경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).<br/>
+ pobi,crong,honux<br/>
+ 시도할 횟수는 몇 회인가요?<br/>
+ 5<br/>
+
+ 실행 결과<br/>
+ pobi : -<br/>
+ crong : -<br/>
+ honux : -<br/>
+
+ pobi : --<br/>
+ crong : -<br/>
+ honux : --<br/>
+
+ pobi : ---<br/>
+ crong : --<br/>
+ honux : ---<br/>
+
+ pobi : ----<br/>
+ crong : ---<br/>
+ honux : ----<br/>
+
+ pobi : -----<br/>
+ crong : ----<br/>
+ honux : -----<br/>
+
+ pobi, honux가 최종 우승했습니다.
