+++ src/main/kotlin/calculator/Main.kt
@@ -1,6 +1,4 @@
-package step2.main
-
-import step2.calculator.Calculator
+package calculator
 
 fun main() {
     val result = Calculator.calculate(readln())
