+++ src/main/kotlin/racing_car/ui/InputView.kt
@@ -1,4 +1,4 @@
-package step3.racing_car.ui
+package racing_car.ui
 
 private const val INPUT_CAR_DESCRIPTION_VIEW = "자동차 대수는 몇 대인가요?\n"
 private const val INPUT_ROUND_DESCRIPTION_VIEW = "시도할 횟수는 몇 회인가요?\n"
